[refactor] 調整漢堡選單的相關函式擺放位置
